requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
Flask==1.1.2
numpy==1.21.2
pandas==1.2.4
Pillow==8.3.2
scikit-image==0.18.3
gunicorn==20.1.0
flask_cors==3.0.10
